+++ lib/components/features/lens-feed/FeedPost.tsx
@@ -1,13 +1,16 @@
-import { ImageMetadata, LinkMetadata, Post, Repost, TextOnlyMetadata } from "@lens-protocol/client";
+import { ImageMetadata, Post, Repost, TextOnlyMetadata } from "@lens-protocol/client";
 import { formatDistanceToNow } from "date-fns";
 
-import { Avatar, Button, toast } from "$lib/components/core";
+import { Avatar, toast } from "$lib/components/core";
 import { getAccountAvatar } from "$lib/utils/lens/utils";
 
 import { FeedPostGallery } from './FeedPostGallery';
 import { PostReaction } from "./PostReaction";
 import { PostRepost } from "./PostRepost";
-import { EventPreview } from "./EventPreview";
+import { PostButton } from "./PostButton";
+import { PostHeader } from "./PostHeader";
+import { PostContent } from "./PostContent";
+import { PostComment } from "./PostComment";
 
 type FeedPostProps = {
   post: Post | Repost;
@@ -52,62 +55,28 @@ export function FeedPost({ post, isComment }: FeedPostProps) {
     <div className="space-y-2">
       <div className="bg-card rounded-md border border-card-border px-4 py-3 space-y-3">
         <div className="flex justify-between">
-          <div className="flex gap-3">
-            <Avatar
-              src={getAccountAvatar(author)}
-              size="xl"
-              rounded="full"
-            />
-            <div className="flex-1">
-              <p>
-                {author.username?.localName}
-              </p>
-              <p className="text-sm text-tertiary">
-                {formatDistanceToNow(new Date(timestamp), { addSuffix: true })}
-              </p>
-            </div>
-          </div>
+          <PostHeader post={rootPost} />
           <div className="flex gap-2">
-            <Button
-              variant="tertiary"
+            <PostButton
+              icon="icon-upload"
               onClick={() => toast.success('Coming soon')}
-              iconLeft="icon-upload"
-              className="rounded-full"
             />
-            <Button
-              variant="tertiary"
+            <PostButton
+              icon="icon-more-vert"
               onClick={() => toast.success('Coming soon')}
-              iconLeft="icon-more-vert"
-              className="rounded-full"
             />
           </div>
         </div>
-        <p className="text-secondary whitespace-pre-line">{(metadata as TextOnlyMetadata).content}</p>
-        {(metadata as ImageMetadata).attachments?.length > 0 && (
-          <FeedPostGallery attachments={(metadata as ImageMetadata).attachments.map(({ item }) => item)} />
-        )}
-        {
-          (metadata as LinkMetadata).sharingLink && (
-            <EventPreview url={(metadata as LinkMetadata).sharingLink} />
-          )
-        }
+        <PostContent post={rootPost} />
         <div className="flex justify-between">
-          <div className="flex gap-2">
+          <div className="flex gap-4 sm:gap-2">
             <PostReaction post={rootPost} />
-            <Button
-              variant="tertiary"
-              iconLeft="icon-chat"
-              className="rounded-full"
-            >
-              {rootPost.stats.comments}
-            </Button>
+            <PostComment post={rootPost} />
             <PostRepost post={rootPost} />
           </div>
-          <Button
-            variant="tertiary"
+          <PostButton
+            icon="icon-share"
             onClick={() => toast.success('Coming soon')}
-            iconLeft="icon-share"
-            className="rounded-full"
           />
         </div>
       </div>