+++ lib/components/features/lens-feed/PostComposer.tsx
@@ -1,4 +1,4 @@
-import { useState, useRef, useEffect } from "react";
+import { useState } from "react";
 import { useAtomValue } from "jotai";
 
 import { Avatar, Button, modal, toast } from "$lib/components/core";
@@ -6,10 +6,13 @@ import { MediaFile, uploadFiles } from "$lib/utils/file";
 import { generatePostMetadata, getAccountAvatar } from "$lib/utils/lens/utils";
 import { accountAtom } from "$lib/jotai";
 import { randomUserImage } from "$lib/utils/user";
+import { useLensAuth } from "$lib/hooks/useLens";
 
 import { ImageInput } from "./ImageInput";
 import { AddEventModal } from "./AddEventModal";
 import { EventPreview } from "./EventPreview";
+import { PostTextarea } from "./PostTextarea";
+import { FileInput } from "./FileInput";
 
 type PostComposerProps = {
   placeholder?: string;
@@ -21,22 +24,15 @@ export function PostComposer({ placeholder, onPost }: PostComposerProps) {
 
   const [value, setValue] = useState('');
   const [isActive, setIsActive] = useState(false);
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
 
   const [files, setFiles] = useState<File[]>([]);
-  const inputRef = useRef<HTMLInputElement>(null);
   const [isUploading, setIsUploading] = useState(false);
 
   const [sharingLink, setSharingLink] = useState<string | undefined>(undefined);
 
   const [isLoading, setIsLoading] = useState(false);
 
-  useEffect(() => {
-    const textarea = textareaRef.current;
-    if (!textarea) return;
-    textarea.style.height = 'auto';
-    textarea.style.height = `${textarea.scrollHeight}px`;
-  }, [value]);
+  const handleLensAuth = useLensAuth();
 
   const handlePost = async () => {
     const content = value.trim();
@@ -68,15 +64,12 @@ export function PostComposer({ placeholder, onPost }: PostComposerProps) {
       <Avatar src={account ? getAccountAvatar(account) : randomUserImage()} size="xl" rounded="full" />
 
       <div className="space-y-4 flex-1">
-        <textarea
-          ref={textareaRef}
+        <PostTextarea
           value={value}
-          onChange={e => setValue(e.target.value)}
-          placeholder={account ? placeholder || `What's on your mind, ${account?.username?.localName}?` : 'Please connect your wallet to post'}
-          className="w-full bg-transparent border-none outline-none font-medium text-lg mt-2 placeholder-quaternary resize-none overflow-hidden min-h-[24px] max-h-[200px]"
+          setValue={setValue}
+          placeholder={placeholder || (account ? `What's on your mind, ${account?.username?.localName}?` : `What's on your mind?`)}
           onFocus={() => setIsActive(true)}
-          rows={1}
-          disabled={!account}
+          className="mt-2"
         />
 
         {
@@ -101,7 +94,20 @@ export function PostComposer({ placeholder, onPost }: PostComposerProps) {
         {isActive && (
           <div className="flex items-center justify-between">
             <div className="flex gap-2">
-              <Button icon="icon-image" variant="tertiary" className="rounded-full" onClick={() => inputRef.current?.click()} />
+              <FileInput
+                onChange={setFiles}
+                accept="image/*"
+                multiple
+              >
+                {open => (
+                  <Button
+                    icon="icon-image"
+                    variant="tertiary"
+                    className="rounded-full"
+                    onClick={open}
+                  />
+                )}
+              </FileInput>
               <Button
                 icon="icon-celebration"
                 variant="tertiary"
@@ -111,6 +117,7 @@ export function PostComposer({ placeholder, onPost }: PostComposerProps) {
                     props: {
                       onConfirm: link => setSharingLink(link),
                     },
+                    dismissible: true
                   });
                 }}
               />
@@ -119,7 +126,7 @@ export function PostComposer({ placeholder, onPost }: PostComposerProps) {
               variant="primary"
               className="rounded-full"
               disabled={!value.trim() || isLoading}
-              onClick={handlePost}
+              onClick={() => handleLensAuth(handlePost)}
               loading={isLoading || isUploading}
             >
               Post
@@ -127,15 +134,6 @@ export function PostComposer({ placeholder, onPost }: PostComposerProps) {
           </div>
         )}
       </div>
-
-      <input
-        ref={inputRef}
-        type="file"
-        multiple
-        className="hidden"
-        accept="image/*"
-        onChange={e => setFiles(e.target.files ? Array.from(e.target.files) : [])}
-      />
     </div>
   );
 }