+++ lib/components/features/theme-builder/generator.tsx
@@ -1,10 +1,30 @@
 'use client';
+import React from 'react';
 import { generateCssVariables } from '$lib/utils/fetchers';
 import clsx from 'clsx';
 import { ShaderGradient } from './shader';
 import { ThemeValues } from './store';
+import { EmojiAnimate } from './emoji';
 
 export function ThemeGenerator({ data }: { data: ThemeValues }) {
+  const [mode, setMode] = React.useState(data.config.mode);
+
+  React.useEffect(() => {
+    if (data.config.mode === 'auto' && window.matchMedia) {
+      const isDarkMode = window.matchMedia('(prefers-color-scheme: dark)').matches;
+      if (isDarkMode) setMode('dark');
+      else setMode('light');
+
+      window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', (event) => {
+        const isDarkMode = event.matches;
+        if (data.config.mode === 'auto') {
+          if (isDarkMode) setMode('dark');
+          else setMode('light');
+        }
+      });
+    }
+  }, [data.config.mode]);
+
   return (
     <>
       {data?.variables && (
@@ -28,12 +48,30 @@ export function ThemeGenerator({ data }: { data: ThemeValues }) {
             'background',
             data?.theme,
             data.config.name,
+            // allow bg for emoji
             data?.config?.color,
             data?.config?.class,
-            data.config.mode,
+            mode,
           )}
         >
-          {data?.theme === 'shader' && <ShaderGradient mode={data.config.mode} />}
+          {data?.theme === 'shader' && <ShaderGradient mode={mode} />}
+
+          {data.config?.effect?.type === 'video' && data?.config?.effect?.url && (
+            <video
+              key={data.config.name}
+              className="min-w-full min-h-full fixed inset-0"
+              autoPlay
+              loop
+              playsInline
+              muted
+            >
+              <source src={data?.config?.effect?.url} type="video/webm"></source>
+            </video>
+          )}
+
+          {data.config.effect?.type === 'float' && (
+            <EmojiAnimate key={data.config.effect.emoji} emoji={data.config.effect.emoji} />
+          )}
         </div>
       )}
     </>